depositary multithreading, consistent hashing, fixes.

+++ depositary/src/main/java/com/apps/depositary/persistance/repository/DepositRepository.java
@@ -7,6 +7,7 @@ import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Repository
@@ -14,6 +15,6 @@ public interface DepositRepository extends JpaRepository<Deposit, UUID> {
 
     @Modifying
     @Query("update Deposit d set d.fillPrice = ?1, d.quantity = ?2 where d.uuid = ?3")
-    void updateDeposit(Double fillPrice, Integer quantity, UUID id);
+    void updateDeposit(BigDecimal fillPrice, Integer quantity, UUID id);
 
 }
